backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
ebooklib==0.18
Pillow==10.2.0
python-multipart==0.0.6
pydantic==2.5.3
python-dotenv==1.0.0
